Move `AtomicAppHandler` to `snow` pkg
